refactor(eager_query): move to models module, via query_options() class methods
